remove early_stopping_config
