--- clinicadl/optim/early_stopping/__init__.py
@@ -1,2 +1,0 @@
-from .config import EarlyStoppingConfig
-from .early_stopper import EarlyStopping
